Disable shadow in Editor as Mantine Paper shadows were propagating
A <Paper> component used as some distant parent of <Editor> could
change the shadow of the editor, making it often have a double
shadow (one for the outer paper, one for the editor).

This is apparently intended behaviour of Mantine, see:
https://github.com/mantinedev/mantine/issues/4930

We explicitly disable shadow on Editor component so it looks flush.